put dx in example repo
