+++ core/.changeset/silver-jokes-switch.md
@@ -1,0 +1,5 @@
+---
+"@crate.ai/discogs-sdk": patch
+---
+
+removed DX from core
