+++ core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@crate.ai/discogs-sdk",
-  "version": "1.3.10",
+  "version": "1.3.11",
   "description": "an SDK for the Discogs API",
   "repository": {
     "type": "git",
@@ -42,10 +42,5 @@
     "@types/node": "^18.19.33",
     "tsup": "^8.0.2",
     "typescript": "^4.4.4"
-  },
-  "dependencies": {
-    "@clack/prompts": "^0.7.0",
-    "open": "^10.1.0",
-    "picocolors": "^1.0.1"
   }
 }
