+++ example/package.json
@@ -12,7 +12,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@crate.ai/discogs-sdk": "file:../core"
+    "@crate.ai/discogs-sdk": "file:../core",
+    "dotenv": "^16.4.5",
+    "@clack/prompts": "^0.7.0",
+    "picocolors": "^1.0.1"
   },
   "devDependencies": {
     "@types/node": "^20.12.12",
